+++ .template-lintrc.js
@@ -916,6 +916,12 @@ module.exports = {
       ]
     },
     {
+      "moduleId": "tests/dummy/app/templates/components-examples/flexberry-groupedit/groupedit-with-multiselect-edit",
+      "only": [
+        "simple-unless"
+      ]
+    },
+    {
       "moduleId": "tests/dummy/app/templates/components-examples/flexberry-groupedit/ember-flexberry-dummy-suggestion-edit-readonly-columns-by-configurate-row-example",
       "only": [
         "simple-unless"
@@ -1072,6 +1078,12 @@ module.exports = {
       ]
     },
     {
+      "moduleId": "tests/dummy/app/templates/components-examples/flexberry-objectlistview/toolbar-custom-components-example",
+      "only": [
+        "eol-last"
+      ]
+    },
+    {
       "moduleId": "tests/dummy/app/templates/components-examples/flexberry-objectlistview/list-on-editform",
       "only": [
         "block-indentation",
@@ -1196,12 +1208,6 @@ module.exports = {
         "require-button-type",
         "require-valid-alt-text",
         "no-bare-strings"
-      ]
-    },
-    {
-      "moduleId": "tests/dummy/app/templates/integration-examples/edit-form/validation",
-      "only": [
-        "block-indentation"
       ]
     },
     {