+++ addon/controllers/lookup-dialog.js
@@ -5,7 +5,8 @@
 import { inject as service } from '@ember/service';
 import { isBlank } from '@ember/utils';
 import { deprecate } from '@ember/debug';
-import { observer } from '@ember/object';
+import { observer, get } from '@ember/object';
+import { scheduleOnce } from '@ember/runloop';
 import ListFormController from '../controllers/list-form';
 import SortableRouteMixin from '../mixins/sortable-route';
 import PredicateFromFiltersMixin from '../mixins/predicate-from-filters';
@@ -35,6 +36,12 @@ export default ListFormController.extend(SortableRouteMixin, PredicateFromFilter
     @type String
   */
   title: undefined,
+
+  /**
+    Flag which indicates that several rows were selected.
+    Shows whether multiselect in groupedit is available.
+  */
+  isPullUpActive: false,
 
   /**
     Current lookup selected record.
@@ -120,6 +127,24 @@ export default ListFormController.extend(SortableRouteMixin, PredicateFromFilter
   */
   lookupEventsService: service('lookup-events'),
 
+  init() {
+    this._super(...arguments);
+
+    this.get('objectlistviewEventsService').on('olvRowSelected', this, this._rowSelected);
+  },
+
+  willDestroy() {
+    this.get('objectlistviewEventsService').off('olvRowSelected', this, this._rowSelected);
+
+    this._super(...arguments);
+  },
+
+  _rowSelected() {
+    const multi = this.get('objectlistviewEventsService').getMultiSelectedRecords(this.get('folvComponentName'));
+    let isPullUpActive = multi && multi.size;
+    this.set('isPullUpActive', isPullUpActive);
+  },
+
   actions: {
     /**
       Handlers OLV row click, Save selected row to object master property and close modal window.
@@ -141,6 +166,40 @@ export default ListFormController.extend(SortableRouteMixin, PredicateFromFilter
     createdModalDialog(modalDialog) {
       this.set('_openedModalDialog', modalDialog);
       this.get('lookupEventsService').lookupDialogOnVisibleTrigger(this.get('componentName'), modalDialog);
+    },
+
+    removeModalDialog() {
+      scheduleOnce('afterRender', () => this.get('objectlistviewEventsService').clearMultiSelectedRecords());
+      return true;
+    },
+
+    pullUpLookupValues() {
+      const currentLookupRow = this.get('currentLookupRow');
+      let selected = this.get('objectlistviewEventsService').getMultiSelectedRecords(this.get('folvComponentName'));
+      const contextModel = this.get('reloadContext.model');
+      const lookupModelName = this.get('saveTo.model.constructor.modelName');
+      const relName = get(contextModel.constructor, 'relationships').get(lookupModelName)[0].name;
+      let detail = contextModel.get(relName);
+      if (selected) {
+        if (currentLookupRow) {
+          selected.forEach(({ data }) => detail.pushObject(this.store.createRecord(lookupModelName, {
+            [this.saveTo.propName]: data
+          })));
+        } else {
+          let rowValueToUpdate = [];
+          selected.forEach(({ data }) => rowValueToUpdate.push(data) );
+          this._selectMaster(rowValueToUpdate[0])
+
+          for (let i = 1; i < rowValueToUpdate.length; i++) {
+            detail.pushObject(this.store.createRecord(lookupModelName, {
+              [this.saveTo.propName]: rowValueToUpdate[i]
+            }))
+          }
+        }
+      }
+  
+      this.get('objectlistviewEventsService').clearMultiSelectedRecords();
+      this._closeModalDialog();
     },
 
     /**
@@ -206,7 +265,10 @@ export default ListFormController.extend(SortableRouteMixin, PredicateFromFilter
         currentLookupRow: this.get('currentLookupRow'),
         customPropertiesData: this.get('customPropertiesData'),
         componentName: this.get('componentName'),
-        folvComponentName: this.get('folvComponentName')
+        folvComponentName: this.get('folvComponentName'),
+        modalDialogSettings: {
+          lookupSettings: this.get('modalDialogSettings.lookupSettings'),
+        }
       };
 
       if (reloadData.customPropertiesData) {