+++ CHANGELOG.md
@@ -3,6 +3,9 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+* The `flexberry-multiple-lookup` component:
+    * Add `chooseComponentVisibility` property.
 
 ## [3.8.0] - 2023-04-25
 ### Added